+++ packages/core/client/Alert.js
@@ -1,2 +1,4 @@
 
 Kepler.Alert = sAlert;
+
+//TODO implements more methods