+++ packages/i18n/i18n/de.js
@@ -71,6 +71,9 @@ i18n.map('de', {
 	btn_users: "Leute",
 	btn_settings: "Einstellungen",
 	btn_addnew: "Neue hinzufügen",
+
+	btn_nearby: "in der Nähe",
+	
 	btn_verify: "Überprüfen",
 	btn_verified: "Überprüft",
 	btn_notverified: "Kein Überprüft",