+++ packages/i18n/i18n/sv.js
@@ -73,6 +73,8 @@ i18n.map('sv', {
 	btn_addnew: "Lägg till ny",
 	btn_accept: "Acceptera",
 	btn_close: "Stänga",
+	btn_nearby: "Nearby",
+	
 	btn_verify: "Verifiera",
 	btn_verified: "Verifierad",
 	btn_notverified: "Inga Verifierad",