+++ packages/lib/package.js
@@ -108,7 +108,7 @@ Package.onUse(function(api) {
     'config/leaflet.js',
     'config/Router.js',
     'config/underscore.js',
-    'config/sAlert.js',
+    //'config/sAlert.js',
     'config/UserPresence.js'
   ]);